Move tests under __tests__
