+++ packages/head-up/src/components/Cell/__tests__/CellSettings.test.js
@@ -1,5 +1,5 @@
 import { shallowMount } from '@vue/test-utils';
-import CellSettings from './CellSettings';
+import CellSettings from '../CellSettings';
 
 test('default render', () => {
   const wrapper = shallowMount(CellSettings);
