+++ packages/head-up/src/components/HeadUp/__tests__/HeadUp.test.js
@@ -4,14 +4,14 @@ import ally from 'ally.js';
 import ShortKey from 'vue-shortkey';
 import Chance from 'chance';
 import { createLocalVue, shallowMount } from '@vue/test-utils';
-import store, { initialState } from '../../store';
-import Board from '../Board';
-import Cell from '../Cell';
-import Sidebar from '../Sidebar';
-import ModalDialogue from '../ModalDialogue';
-import VSwitchToggle from '../VSwitchToggle';
-import HeadUp from './HeadUp';
-import HeadUpBoards from './HeadUpBoards';
+import store, { initialState } from '../../../store';
+import Board from '../../Board';
+import Cell from '../../Cell';
+import Sidebar from '../../Sidebar';
+import ModalDialogue from '../../ModalDialogue';
+import VSwitchToggle from '../../VSwitchToggle';
+import HeadUp from '../HeadUp';
+import HeadUpBoards from '../HeadUpBoards';
 
 const localVue = createLocalVue();
 localVue.use(Vuex);
