+++ packages/head-up/src/components/Sidebar/__tests__/SidebarButton.test.js
@@ -1,5 +1,5 @@
 import { shallowMount } from '@vue/test-utils';
-import SidebarButton from './SidebarButton';
+import SidebarButton from '../SidebarButton';
 
 test('render default', () => {
   const wrapper = shallowMount(SidebarButton, {
